fix(function): return statement execution bug
There was a bug when return statement was executed in recursive function
calls. It was because the fallback to enclosing environment was not being
executed in the executeblock for return statement due to error propagation.

It was fixed by setting the interpreter's environment back to the enclosing
environment in the call function.